+++ NEWS.md
@@ -1,6 +1,6 @@
 # dplyr 0.7.1.9000
 
-* `distinct()` now throws an error when used on unknown columns (#2867).
+* `distinct()` now throws an error when used on unknown columns (#2867, @foo-bar-baz-qux).
 
 # dplyr 0.7.1